consistent splitting binding
there are two key bindings to split in each direction (horizontally and
vertically). One to split the current pane, and another to split the
whole width or height of the screen.

Before this commit the bindings were inconsistent. to split the currnt
pane the binding was with shift horizontally `|` (which is shift `\`),
and without shift vertically `-`.

Now both pane splitting bindings do not require shift: `\` and `-`.